Split the companies house api into two parts, the search api and the auth api
