+++ src/main/scala/services/CompanyAuthService.scala
@@ -18,33 +18,13 @@
 package services
 
 import com.google.inject.ImplementedBy
-import com.wellfactored.playbindings.ValueClassFormats
 import models.CompaniesHouseId
-import play.api.libs.json.Json
 import services.companiesHouse.CompaniesHouseAuth
 
 import scala.concurrent.Future
 
-case class CompanySummary(company_number: CompaniesHouseId, title: String, address_snippet: String)
-
-case class CompanyDetail(company_number: CompaniesHouseId, company_name: String)
-
-object CompanyDetail extends ValueClassFormats {
-  implicit val fmt = Json.format[CompanyDetail]
-}
-
-case class ResultsPage(
-                        page_number: Int,
-                        start_index: Int,
-                        items_per_page: Int,
-                        total_results: Int,
-                        items: List[CompanySummary]
-                      )
-
-
 @ImplementedBy(classOf[CompaniesHouseAuth])
 trait CompanyAuthService {
-
   def isInScope(companiesHouseIdentifier: CompaniesHouseId, oAuthToken: OAuthToken): Future[Boolean]
 
   def emailAddress(oAuthToken: OAuthToken): Future[Option[String]]
