+++ src/main/twirl/views/shared/_companySummary.scala.html
@@ -1,4 +1,4 @@
-@(result: services.CompanySummary, reportCount:Int, link: String)
+@(result: services.CompanySearchResult, reportCount:Int, link: String)
 <li class="searchresult">
     <h2 class="heading-medium">
         <a href=@link>@result.title</a>
